Safari-609.1.11.1/LayoutTests/http/tests/eventsource/eventsource-url-attribute-expected.txt

Both .URL and .url should work (for compatibility reasons).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS source.URL is "http://127.0.0.1:8000/eventsource/resources/event-stream.php"
PASS source.url is "http://127.0.0.1:8000/eventsource/resources/event-stream.php"
PASS source.URL === source.url is true
PASS successfullyParsed is true

TEST COMPLETE
